İced river's and weak bridge's specialities added.

+++ Hackathon/Assets/Scripts/PlayerController.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class PlayerController : MonoBehaviour
 {
-    public int maxMoveDistance = 1;
+    public float maxMoveDistance = 1f;
     public float movementSpeed = 5f;
     public int posX, posY;
